perf(jira integration): parallelize API calls for improved performance; refactor: split jira.controller.ts into separate files for better maintainability
