+++ server/src/controller/backend/jira.controller/index.ts
@@ -1,0 +1,50 @@
+// Central export file for Jira controllers
+// This file re-exports all modular Jira controller endpoints for use in routes.
+
+export {
+  getJiraProjects,
+  selectJiraProject,
+  getSelectedJiraProjectId,
+  GetJiraProjectsQueryParams,
+  GetJiraProjectsResponse,
+  JiraProject,
+  SelectJiraProjectBody,
+  SelectJiraProjectResponse,
+  GetSelectedJiraProjectIdQueryParams,
+  GetSelectedJiraProjectIdResponse
+} from './projects';
+
+export {
+  getJiraTicketsForProject,
+  getJiraTicketsForSprint,
+  getJiraIssue,
+  editJiraIssue,
+  GetJiraTicketsQueryParams,
+  GetJiraTicketsResponse,
+  JiraTicket,
+  GetJiraTicketsForSprintQueryParams,
+  GetJiraTicketsForSprintResponse,
+  GetJiraIssueQueryParams,
+  GetJiraIssueResponse,
+  DetailedJiraIssue,
+  EditJiraIssueBody,
+  EditJiraIssueResponse
+} from './tickets';
+
+export {
+  getAllSprintsForProject,
+  getActiveSprintsPerProject,
+  GetSprintsQueryParams,
+  GetSprintsResponse,
+  GetActiveSprintsResponse,
+  JiraSprint
+} from './sprints';
+
+export {
+  getJiraBoardForProject,
+  GetJiraBoardQueryParams,
+  GetJiraBoardResponse,
+  JiraBoard
+} from './boards';
+
+
